requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
faiss-cpu
chromadb
pypdf
pdfplumber
tavily-python
python-dotenv
streamlit
SpeechRecognition
pyttsx3
pydantic
dateparser
